Replace threading w/ multiprocessing + fix bugs
